move post tests to actions package
